# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Covalence API",
        "description": "Hybrid GraphRAG knowledge engine API",
        "license": {
            "name": "MIT",
            "identifier": "MIT"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/sources": {
            "get": {
                "tags": [
                    "sources"
                ],
                "summary": "List sources with pagination.",
                "operationId": "list_sources",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of results (default 20).",
                        "required": false,
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "format": "int64"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for pagination (default 0).",
                        "required": false,
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SourceResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "sources"
                ],
                "summary": "Ingest a new source.",
                "operationId": "create_source",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSourceRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Source created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateSourceResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/sources/{id}": {
            "get": {
                "tags": [
                    "sources"
                ],
                "summary": "Get a source by ID.",
                "operationId": "get_source",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Source ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SourceResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Source not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "sources"
                ],
                "summary": "Delete a source and its chunks.",
                "operationId": "delete_source",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Source ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteSourceResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sources/{id}/chunks": {
            "get": {
                "tags": [
                    "sources"
                ],
                "summary": "Get chunks for a source.",
                "operationId": "get_source_chunks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Source ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chunks for source",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChunkResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ChunkResponse": {
                "type": "object",
                "description": "Response for a chunk entity.",
                "required": [
                    "id",
                    "source_id",
                    "level",
                    "ordinal",
                    "content",
                    "token_count"
                ],
                "properties": {
                    "content": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "level": {
                        "type": "string"
                    },
                    "ordinal": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "source_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "token_count": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "CreateSourceRequest": {
                "type": "object",
                "description": "Request body for source ingestion.",
                "required": [
                    "content",
                    "source_type"
                ],
                "properties": {
                    "content": {
                        "type": "string",
                        "description": "Base64-encoded content bytes."
                    },
                    "metadata": {
                        "description": "Optional metadata."
                    },
                    "mime": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "MIME type of the content (e.g. \"text/markdown\", \"text/plain\").\nDefaults to \"text/plain\" if omitted."
                    },
                    "source_type": {
                        "type": "string",
                        "description": "Source type (document, web_page, conversation, code, api, manual,\ntool_output, observation)."
                    },
                    "uri": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Optional URI of the original material."
                    }
                }
            },
            "CreateSourceResponse": {
                "type": "object",
                "description": "Response after successful source creation.",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID of the created (or deduplicated) source."
                    }
                }
            },
            "DeleteSourceResponse": {
                "type": "object",
                "description": "Response for source deletion.",
                "required": [
                    "deleted",
                    "chunks_deleted"
                ],
                "properties": {
                    "chunks_deleted": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "deleted": {
                        "type": "boolean"
                    }
                }
            },
...
            "SourceResponse": {
                "type": "object",
                "description": "Response for a source entity.",
                "required": [
                    "id",
                    "source_type",
                    "ingested_at",
                    "reliability_score",
                    "clearance_level",
                    "content_version"
                ],
                "properties": {
                    "author": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "clearance_level": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content_version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "ingested_at": {
                        "type": "string"
                    },
                    "reliability_score": {
                        "type": "number",
                        "format": "double"
                    },
                    "source_type": {
                        "type": "string"
                    },
                    "title": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "uri": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
        }
    },
...
}